+++ changelog.md
@@ -5,6 +5,7 @@
 ### Major
 
 - Updated Node.js support to `^14.17.0 || ^16.0.0 || >= 18.0.0`.
+- Updated the [`react`](https://npm.im/react) and [`react-dom`](https://npm.im/react-dom) peer dependencies to `^18.0.0`.
 - Updated dependencies, some of which require newer Node.js versions than previously supported.
 - Use the `node:` URL scheme for Node.js builtin module imports in tests.
 
@@ -18,6 +19,7 @@
 - Updated GitHub Actions CI config:
   - Run tests with Node.js v14, v16, v18.
 - Removed the now redundant `not IE > 0` from the Browserslist query.
+- Updated `react-dom/server` imports to suit React v18 in tests.
 - Corrected “Markdown” capitalization in documentation.
 - Added a JSDoc description to the React component `Margin`.